Remove LLVM logic from the generic Device interface
