+++ taichi/backends/device.cpp
@@ -140,15 +140,6 @@ void Device::print_all_cap() const {
   }
 }
 
-uint64_t *Device::allocate_llvm_runtime_memory_jit(
-    const LlvmRuntimeAllocParams &params) {
-  params.runtime_jit->call<void *, std::size_t, std::size_t>(
-      "runtime_memory_allocate_aligned", params.runtime, params.size,
-      taichi_page_size);
-  return taichi_union_cast_with_different_sizes<uint64_t *>(fetch_result_uint64(
-      taichi_result_buffer_runtime_query_id, params.result_buffer));
-}
-
 void GraphicsDevice::image_transition(DeviceAllocation img,
                                       ImageLayout old_layout,
                                       ImageLayout new_layout) {
